+++ Changelog.markdown
@@ -10,6 +10,10 @@
 
 ----
 
+#### Version 0.6.2 (2013-07-12):
+
+* __upd__:  `modes.Modes()` renamed to `modes.Parser()`
+
 #### Version 0.2.2 (2013-06-29):
 
 * __upd__:  if option requires no argument `None` is returned instead of raising `TypeError`